State: Move site deletion notices away from middleware
